ItemUtils and InventoryUtils where partially overlapping. Move item stuff to ItemUtils.
