+++ common/src/main/java/com/wynntils/utils/wynn/ItemUtils.java
@@ -5,8 +5,10 @@
 package com.wynntils.utils.wynn;
 
 import com.wynntils.core.components.Models;
+import com.wynntils.core.text.StyledText;
 import com.wynntils.models.gear.type.GearTier;
-import com.wynntils.models.items.items.game.GearItem;
+import com.wynntils.models.items.properties.GearTypeItemProperty;
+import java.util.List;
 import java.util.Optional;
 import java.util.regex.Pattern;
 import net.minecraft.ChatFormatting;
@@ -20,10 +22,31 @@ public final class ItemUtils {
             Pattern.compile("(Normal|Set|Unique|Rare|Legendary|Fabled|Mythic)( Raid)? (Item|Reward).*");
 
     public static boolean isWeapon(ItemStack itemStack) {
-        Optional<GearItem> gearItemOpt = Models.Item.asWynnItem(itemStack, GearItem.class);
+        Optional<GearTypeItemProperty> gearItemOpt =
+                Models.Item.asWynnItemPropery(itemStack, GearTypeItemProperty.class);
         if (gearItemOpt.isEmpty()) return false;
 
         return gearItemOpt.get().getGearType().isWeapon();
+    }
+
+    public static StyledText getItemName(ItemStack itemStack) {
+        return StyledText.fromComponent(itemStack.getHoverName());
+    }
+
+    public static boolean isItemListsEqual(List<ItemStack> firstItems, List<ItemStack> secondItems) {
+        if (firstItems.size() != secondItems.size()) return false;
+
+        for (int i = 0; i < firstItems.size(); i++) {
+            ItemStack newItem = firstItems.get(i);
+            ItemStack oldItem = secondItems.get(i);
+            if (!newItem.getItem().equals(oldItem.getItem())
+                    || newItem.getDamageValue() != oldItem.getDamageValue()
+                    || newItem.getCount() != oldItem.getCount()
+                    || !ItemStack.tagMatches(oldItem, newItem)) {
+                return false;
+            }
+        }
+        return true;
     }
 
     public static MutableComponent getNonGearDescription(ItemStack itemStack, String gearName) {
